+++ web/src/modules/transactions/Table/Pagination/style.ts
