+++ web/webpack.dev.ts
@@ -23,6 +23,7 @@ const config: webpack.Configuration = {
       '@Globals': path.resolve(__dirname, 'src/globals'),
       '@Generated': path.resolve(__dirname, 'src/generated'),
       '@Lib': path.resolve(__dirname, 'src/lib'),
+      '@Modules': path.resolve(__dirname, 'src/modules'),
     },
   },
   ignoreWarnings: [/Failed to parse source map/],
